refactor(237): update teacher architecture for tests
